Algo

+++ src/classes/business/Product.java
@@ -8,9 +8,9 @@ public class Product{
     private int id;
     private String name;
     private double price;
-    private int quantity;
+    private double quantity;
 
-    public Product(int id, String name, double price, int quantity) {
+    public Product(int id, String name, double price, double quantity) {
         this.id = id;
         this.name = name;
         this.price = price;
@@ -29,7 +29,7 @@ public class Product{
         return this.price;
     }
 
-    public int getQuantity() {
+    public double getQuantity() {
         return this.quantity;
     }
